% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_fast_tsne_path}
\alias{get_fast_tsne_path}
\title{Get FIt-SNE executable path}
\usage{
get_fast_tsne_path()
}
\value{
Character string with the path to the fast_tsne executable
}
\description{
Detects and returns the correct path to the FIt-SNE (Fast Fourier Transform
accelerated t-SNE) executable based on the current system. Checks common
installation locations for HPC, macOS, and Ubuntu systems.
}
\details{
The function checks the following paths in order:
\enumerate{
\item HPC: \verb{/services/tools/fit-sne/1.2.1/bin/fast_tsne}
\item macOS: \verb{/usr/local/FIt-SNE/bin/fast_tsne}
\item Ubuntu: \verb{~/FIt-SNE/bin/fast_tsne}
}

If none of these paths exist, the function stops with an error.
}
\examples{
\dontrun{
# Get the path and use with Seurat
fast_tsne_path <- get_fast_tsne_path()
seurat_obj <- RunTSNE(seurat_obj,
                      tsne.method = "FIt-SNE",
                      fast_tsne_path = fast_tsne_path)
}
}
