% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{jaccard_similarity}
\alias{jaccard_similarity}
\title{Calculate Jaccard similarity between two sets}
\usage{
jaccard_similarity(set1, set2)
}
\arguments{
\item{set1}{First set (vector)}

\item{set2}{Second set (vector)}
}
\value{
Numeric Jaccard similarity coefficient (0-1)
}
\description{
Calculate Jaccard similarity between two sets
}
\keyword{internal}
